requirements.txt
streamlit==1.50.0
anthropic==0.84.0
google-genai==1.33.0
fastapi==0.135.1
uvicorn==0.41.0
PyMuPDF==1.27.1
Pillow==11.3.0
PyYAML==6.0.3
python-dotenv==1.2.2
SQLAlchemy==2.0.48
alembic==1.18.4
pytest==9.0.2
httpx==0.28.1
PyJWT==2.12.0
bcrypt==5.0.0
psycopg2-binary==2.9.11
cryptography==46.0.5
pyotp==2.9.0
# Optional: Redis for multi-worker rate limiting
# redis==5.2.1
